requirements.txt
aiogram
python-dotenv
asyncpg
psycopg2-binary
